fix(openai)!: align tool schemas, requests and resource handling
Use reference provider-tool schemas and normalize options, namespaces and scalar tool outputs. Align usage, resource paths and streamed uploads with regression fixtures and reviewed snapshots.

BREAKING CHANGE: OpenAI and compatible adapters preserve supplied schemas instead of applying OpenAiStrict automatically. Provider-tool factories enforce reference input schemas and defaults.

+++ crates/providers/ferrin-openai-compatible/src/embedding.rs
@@ -62,7 +62,7 @@ struct EmbeddingResponse {
 
 #[derive(Debug, Deserialize)]
 struct EmbeddingData {
-    embedding: Vec<f32>,
+    embedding: Vec<f64>,
 }
 
 #[derive(Debug, Deserialize)]
